Fixed the logic to embed list schema validation in RSpec
