+++ Rakefile
@@ -4,8 +4,9 @@ require 'bundler/gem_tasks'
 require 'rspec/core/rake_task'
 
 require 'google/cloud/bigquery'
-require_relative 'lib/dfe/reference_data/all_lists'
 require 'debug'
+
+require_relative 'lib/dfe/reference_data'
 
 RSpec::Core::RakeTask.new(:spec)
 
@@ -98,12 +99,12 @@ def create_bigquery_table(dataset, table_name, list)
   end
 end
 
-def setup_bigquery_table_metadata(table, table_name, list, version, commit)
-  table.description = ("Automatically populated from dfe-reference-data")
+def setup_bigquery_table_metadata(table, _table_name, _list, version, commit)
+  table.description = ('Automatically populated from dfe-reference-data')
   table.labels = {
     autogenerated: true,
-    dfe_reference_data_version: version.gsub(".","_"),
-    dfe_reference_data_commit_hash: commit,
+    dfe_reference_data_version: version.gsub('.', '_'),
+    dfe_reference_data_commit_hash: commit
   }
 end
 
@@ -129,7 +130,7 @@ MAX_SLEEP = 60
 
 def update_reference_list_into_bigquery_table_with_retries(dataset, table_name, list)
   sleep_time = 1
-  while true
+  loop do
     begin
       table = dataset.table table_name
       if table.exists?
@@ -138,14 +139,12 @@ def update_reference_list_into_bigquery_table_with_retries(dataset, table_name, 
       else
         puts "Table readiness check failed as it doesn't exist yet, sleeping and retrying..."
       end
-    rescue Google::Cloud::NotFoundError => error
-      puts "Table insert attempt failed with #{error}, sleeping and retrying..."
+    rescue Google::Cloud::NotFoundError => e
+      puts "Table insert attempt failed with #{e}, sleeping and retrying..."
     end
     sleep sleep_time
-    sleep_time = sleep_time * 2
-    if sleep_time > MAX_SLEEP
-      sleep_time = MAX_SLEEP
-    end
+    sleep_time *= 2
+    sleep_time = MAX_SLEEP if sleep_time > MAX_SLEEP
   end
 end
 
@@ -179,6 +178,10 @@ task :update_bigquery_tables do
 
   dataset = project.dataset BIGQUERY_DATASET
 
+  # No, rubocop, I do not want to combine these loops because there is an asynchronous process in BigQuery between table creation and table being ready to use, and I wish to start that for all tables before I start trying to use the tables - as retrying and waiting until the first table becomes ready can happen while other tables we will look at later are also becoming ready.
+
+  # rubocop:disable Style/CombinableLoops
+
   BIGQUERY_TABLES.each do |entry|
     (table_name, list) = entry
     puts "Creating #{table_name}..."
@@ -191,16 +194,6 @@ task :update_bigquery_tables do
     puts "Populating #{table_name}..."
     update_reference_list_into_bigquery_table_with_retries(dataset, table_name, list)
   end
-end
 
-## ABS FIXME This is a standin until I can get spec/lib/dfe/reference_data/all_lists_spec.rb to work
-task :validate_lists do
-  DfE::ReferenceData::ALL_LISTS.each do |name, list|
-    puts "Validating #{name}..."
-    errors = list.validate
-    errors.each_entry do |record, error|
-      puts "RECORD: #{record}"
-      puts "HAS ERROR: #{error}"
-    end
-  end
+  # rubocop:enable Style/CombinableLoops
 end
