+++ lib/dfe/reference_data/degrees/grades.rb
@@ -1,3 +1,5 @@
+require_relative '../hardcoded_reference_list'
+
 module DfE
   module ReferenceData
     module Degrees
