+++ lib/dfe/reference_data/degrees/subjects.rb
@@ -1,3 +1,6 @@
+require_relative '../hardcoded_reference_list'
+require_relative '../joined_reference_list'
+
 module DfE
   module ReferenceData
     module Degrees
