+++ spec/lib/dfe/reference_data/degrees/institutions_spec.rb
@@ -1,0 +1,5 @@
+require 'support/schema_validation'
+
+RSpec.describe DfE::ReferenceData::Degrees::INSTITUTIONS do
+  it_should_behave_like 'a list of valid records'
+end
